almost there
everything works and all the pages are connected. For the beginning and ending pages the enter and r keys are not working When i hit enter, it doesn't go to the level 1 page from the beginning page. Same thing happens when i try to restart the game from the final page to go back to beginning page. ALso r doesn't do the reload